Add LicenseExtensions class for license management
Introduces the `LicenseExtensions` static class in the
`MinimalApi.Identity.Licenses.DependencyInjection` namespace.
This class includes constant strings for license-related
endpoints and status messages. It also adds the
`LicenseRegistrationService` method to facilitate the
registration of license services and configure FluentValidation
for the `AssignLicenseValidator`. Necessary using directives
for dependencies are also included.