Remove README.md from package metadata and packing
The `<PackageReadmeFile>` element was removed from the project file, indicating that the README.md file will no longer be included as part of the package metadata. Additionally, the entry for the README.md file in the `<ItemGroup>` was also removed, meaning that the README.md file will not be packed with the project anymore.