Add GitHub Actions workflow for NuGet publishing
This commit introduces a new workflow in `publish_policy_manager.yml` to automate the publishing of the `Identity.Module.PolicyManager` package to NuGet. The workflow triggers on pushes to the `main` branch and changes in the specified source path. It includes steps for code checkout, .NET Core SDK setup, versioning with Nerdbank.GitVersioning, packaging, and publishing the NuGet package. Additionally, it creates a GitHub release with the relevant tag and release name.