Add GitHub Actions workflow for NuGet publishing
This commit introduces a new workflow named "Publish Identity.Module.Results on NuGet" in the `publish_results.yml` file. The workflow is triggered on pushes to the `main` branch and changes in the `src/MinimalApi.Identity.Results/**` path. It sets up environment variables for .NET version, project details, and includes jobs for checking out code, setting up the .NET SDK, managing versioning with Nerdbank.GitVersioning, packaging the project, publishing to NuGet, and creating a GitHub release with the generated version information.