Add commented-out code for service provider and policies
This commit introduces commented-out lines for creating a service provider and an asynchronous scope to access `IAuthPolicyService`. It also includes a line to fetch all policies asynchronously. The `AddAuthorization` method is updated with a commented-out option to add default security options based on the retrieved policies, alongside a call to `AddDefaultSecurityOptions()` without parameters.