Add GitHub Actions workflow for NuGet package publishing
This commit introduces a new workflow in `publish_auth_manager.yml` that triggers on pushes to the `main` branch and changes in the `src/MinimalApi.Identity.AuthManager/**` path. It sets up necessary environment variables and includes steps for checking out the code, configuring the .NET SDK, managing versioning with Nerdbank.GitVersioning, packaging the project, publishing to NuGet, and creating a GitHub release with the generated package version.