Remove CheckCreateUserAsync method from AuthenticationStartupTask
The method `CheckCreateUserAsync` has been removed from the `AuthenticationStartupTask` class. This method was responsible for checking if a user exists by email and creating a new user if they did not exist, along with assigning roles to the user upon successful creation. The removal of this method indicates a potential change in how user creation and role assignment are handled in the application.