Remove obsolete UserLicense deletion check
The commented-out code that checked for the existence of a
`UserLicense` entity before deleting a `License` entity has
been removed. This check, which would have thrown a
`BadRequestException` if a `UserLicense` was found, is no
longer necessary. The updated code focuses on retrieving
and deleting the `License` entity, throwing a
`NotFoundException` if the `License` does not exist.