Add ErrorResponseFormat enum as a temporary placeholder
A new `ErrorResponseFormat` enum was added to the
`MinimalApi.Identity.Core.Enums` namespace with two values:
`Default` and `List`. This enum is intended to define error
response formats. A `TODO` comment was included to indicate
that this enum should eventually be replaced with an equivalent
enum from the `MinimalApi.Identity.Shared.Results.AspNetCore.Http`
namespace. This addition serves as an interim solution.