Add Serilog support for MinIO cloud storage logging
Introduced a `UseSerilogToStorageCloud` extension method in
`SharedExtensions` to enable `IHostBuilder` to configure
Serilog logging with MinIO as a log sink. This method
conditionally writes logs to MinIO based on the presence of
`AccessKey` and `SecretKey` in `MinioOptions`.

Also added a `using` directive for `Microsoft.Extensions.Hosting`
to support `IHostBuilder` functionality.